+++ sgbridgebot/ChatBot.py
@@ -4,6 +4,7 @@ from telegram.ext import Updater, CommandHandler, RegexHandler
 from GameManager import GameManager
 from CommandUtils import CommandUtils
 from ChatHandler import ChatHandler
+import os
 
 class ChatBot(object):
 
@@ -26,16 +27,26 @@ class ChatBot(object):
 
     def __init__(self, token):
         self.updater = Updater(token)
+        self.TOKEN = token
         self.chat_handler = ChatHandler(self.updater.bot)
         self.manager = GameManager(self.chat_handler)
         self.cmd_utils = CommandUtils(self.manager, self.chat_handler)
 
 
-    def start(self):
+    def start(self, poll):
         self.init_command_handlers()
         self.init_regex_handlers()
-        self.updater.start_polling()
-        self.updater.idle()
+
+        if pollhook:
+            self.updater.start_polling()
+            self.updater.idle()
+        else:
+            PORT = int(os.environ.get('PORT', '8443'))
+            updater.start_webhook(listen="0.0.0.0",
+                                  port=PORT,
+                                  url_path=self.TOKEN)
+            updater.bot.set_webhook("https://sgbridgebot.herokuapp.com/" + TOKEN)
+            updater.idle()
 
     def init_regex_handlers(self):
         self.updater.dispatcher.add_handler(RegexHandler('^(PASS|(1|2|3|4|5|6|7)('+"|".join([u'\U00002663',u'\U00002666',u'\U00002764',u'\U00002660','NT'])+'))$', self.cmd_utils.bidding))
@@ -44,6 +55,6 @@ class ChatBot(object):
 
     def init_command_handlers(self):
         self.updater.dispatcher.add_handler(CommandHandler('forcestart', self.cmd_utils.forcestart))
-        self.updater.dispatcher.add_handler(CommandHandler('hello', self.cmd_utils.hello))
         self.updater.dispatcher.add_handler(CommandHandler('join', self.cmd_utils.join))
         self.updater.dispatcher.add_handler(CommandHandler('leave', self.cmd_utils.leave))
+        self.updater.dispatcher.add_handler(CommandHandler('hand', self.cmd_utils.hand))