+++ sgbridgebot/CommandUtils.py
@@ -1,6 +1,7 @@
 from telegram.ext import Updater, CommandHandler
 from BridgeGame import BridgeGame
 from GameManager import GameManager
+from BridgeCard import BridgeCard
 import logging
 
 
@@ -43,11 +44,7 @@ class CommandUtils(object):
             pass
         if (game.state == 0):
             game.start_game()
-
-    # /hello command util
-    def hello(self, bot, update):
-        update.message.reply_text(
-            'Hello {}.'.format(update.message.from_user.first_name))
+        self.manager.update_gamelists()
 
 
     # /join command util
@@ -57,7 +54,7 @@ class CommandUtils(object):
         game = self.manager.join_game(user, update.message.chat)
 
         if game == -1:
-            update.message.reply_text('Could not join: all games are currently full.')
+            update.message.reply_text('Could not join: games are currently full.')
             return
 
         elif game == -2:
@@ -67,7 +64,6 @@ class CommandUtils(object):
         elif not isinstance(game, BridgeGame):
             raise TypeError('GameManager.join_game() did not return a valid BridgeGame object')
 
-        update.message.reply_text('You joined the game.')
         # send broadcast update to all members of the room
         self.chat.player_joined_game(user, game)
 
@@ -89,6 +85,21 @@ class CommandUtils(object):
         # send broadcast update to all members of the room
         self.chat.player_left_game(user, game)
 
+    def hand(self, bot, update):
+        user = update.message.from_user
+        chat_id = update.message.chat_id
+        for g in self.manager.active_games:
+            for p in g.players:
+                if p.id == user.id:
+                    if update.message.chat.type != 'private':
+                        self.chat.ask_private_chat(g)
+                        return
+                    self.chat.display_game_players(chat_id, g)
+                    self.chat.display_hand(p, chat_id)
+
+    '''
+    RegexHandlers
+    '''
 
     def bidding(self, bot, update):
         # handles all text messgaes that have the format NUMBER|SUIT
@@ -106,12 +117,27 @@ class CommandUtils(object):
         user = update.message.from_user
         game = self.manager.find_game(user, update.message.chat_id)
         card_str = update.message.text
+        card_id = self.chat.str_utils.card_str_to_id(card_str)
         if isinstance(game, BridgeGame):
             if game.state == 2 and user.id == game.curr_player().id:
+                # Chose a partner, update game state
                 update.message.reply_text('Partner chosen.')
-                card_id = self.chat.str_utils.card_str_to_id(card_str)
                 game.partner = game.player_holding_card(card_id)
                 self.chat.partner_chosen(game.curr_player(), card_id, game)
+                # update state
+                game.next_turn()
             elif game.state == 3 and user.id == game.curr_player().id:
-                update.message.reply_text('You played a card.')
+                # Played a card, update game state
+                card = BridgeCard(card_id)
+                if card.id in [c.id for c in game.curr_player().hand]:
+                    update.message.reply_text('You played a card.')
+                    game.curr_player().remove_card(card)
+                    game.trick.append(card)
+                    self.chat.card_played(game.curr_player(), card, game)
+                    game.next_turn()
+                else:
+                    update.message.reply_text('Invalid card!')
+                    self.chat.request_card(game.curr_player(), game)
+            else:
+                update.message.reply_text('Not you turn to play a card! Current turn: {}'. format(game.curr_player().disp_name()))
         return