+++ CHANGELOG.md
@@ -6,17 +6,26 @@ Add a section here for every commit that is made, in the following format: major
 
 Build numbers are included in the alpha stage of development, and will be dropped eventually. For now time is in UTC+5.
 
+## 0.0.6-alpha: build 180602/1841
+Upgraded to python 3.6
++ players can now play a full game
++ implemented trick winning logic
++ implemented game winning logic based on contract
++ cards now shown on an inline keyboard - players can select card to play
+- have yet to implement card checking
+- bots do not choose cards well
+
+
 ## 0.0.5-alpha: build 180602/0135
 Implemented partner choice
 + Players can choose their partner using inline keyboard
-+ Fixed card values not displaying properly
+- Bug: card values not displaying properly sometimes
 
 ## 0.0.5-alpha: build 180601/0221
 Fully implemented bidding system
 + bots can now choose their partner
 + bots now randomly choose a bid
 + one pass around ends bidding phase
-
 
 ## 0.0.4-alpha: build 180531/2216
 Partially implemented bidding system